Add convex hull computer trait, start refactor with tests
