+++ src/polygon.rs
@@ -69,7 +69,7 @@ impl Polygon {
         // need something to initiate algorithms in the vertex chain.
         // Could consider only exposing keys and then having polygon gen
         // an anchor
-        self.vertex_map.values().collect::<Vec<_>>()[0]
+        self.vertices()[0]
     }
 
     pub fn num_edges(&self) -> usize {
@@ -93,8 +93,12 @@ impl Polygon {
             .collect::<Vec<Point>>()
     }
 
+    pub fn vertices(&self) -> Vec<&Vertex> {
+        self.vertex_map.values().collect_vec()
+    }
+
     pub fn sorted_vertices(&self) -> Vec<&Vertex> {
-        let mut vertices = self.vertex_map.values().collect_vec();
+        let mut vertices = self.vertices();
         vertices.sort_by(|a, b| a.id.cmp(&b.id));
         vertices
     }
@@ -390,42 +394,6 @@ impl Polygon {
         hull
     }
 
-    pub fn convex_hull_from_gift_wrapping(&self) -> ConvexHull {
-        let mut hull = ConvexHull::default();
-        // Form a horizontal line terminating at lowest point to start
-        let v0 = self.leftmost_lowest_vertex();
-        let mut p = v0.coords.clone();
-        p.x -= 1.0;  // Arbitrary distance
-        let mut current_edge = LineSegment::new(&p, &v0.coords);
-        let mut current_vertex_id = v0.id;
-        hull.add_vertex(current_vertex_id);
-
-        // Perform gift-wrapping, using the previous hull edge as a vector to 
-        // find the point with the least CCW angle w.r.t. the vector. Connect 
-        // that point to the current terminal vertex to form the newest hull 
-        // edge. Repeat until we reach the starting vertex again.
-        loop {
-            let min_angle_vertex_id = self.vertex_map
-                .values()
-                .filter(|v| v.id != current_vertex_id)
-                .min_by_key(|v| OrderedFloat(current_edge.angle_to_point(&v.coords)))
-                .unwrap()
-                .id;
-
-            current_edge = self.get_line_segment(
-                &current_vertex_id, &min_angle_vertex_id
-            ).unwrap();
-            current_vertex_id = min_angle_vertex_id;
-            if current_vertex_id == v0.id {
-                break;
-            } else {
-                hull.add_vertex(current_vertex_id);
-            }
-
-        }
-        hull
-    }
-
     pub fn bounding_box(&self) -> BoundingBox {
         BoundingBox::new(self.min_x(), self.max_x(), self.min_y(), self.max_y())
     }
@@ -447,25 +415,25 @@ impl Polygon {
     }
 
     pub fn leftmost_lowest_vertex(&self) -> &Vertex {
-        let mut vertices = self.vertex_map.values().collect_vec();
+        let mut vertices = self.vertices();
         vertices.sort_by_key(|v| (OrderedFloat(v.coords.y), OrderedFloat(v.coords.x)));
         vertices[0]
     }
 
     pub fn rightmost_lowest_vertex(&self) -> &Vertex {
-        let mut vertices = self.vertex_map.values().collect_vec();
+        let mut vertices = self.vertices();
         vertices.sort_by_key(|v| (OrderedFloat(v.coords.y), OrderedFloat(-v.coords.x)));
         vertices[0]
     }
 
     pub fn lowest_rightmost_vertex(&self) -> &Vertex {
-        let mut vertices = self.vertex_map.values().collect_vec();
+        let mut vertices = self.vertices();
         vertices.sort_by_key(|v| (OrderedFloat(-v.coords.x), OrderedFloat(v.coords.y)));
         vertices[0]
     }
 
     pub fn highest_leftmost_vertex(&self) -> &Vertex {
-        let mut vertices = self.vertex_map.values().collect_vec();
+        let mut vertices = self.vertices();
         vertices.sort_by_key(|v| (OrderedFloat(v.coords.x), OrderedFloat(-v.coords.y)));
         vertices[0]
     }
@@ -575,152 +543,14 @@ impl Polygon {
 #[cfg(test)]
 mod tests {
     use crate::F64_ASSERT_PRECISION;
-    use crate::util::load_polygon;
+    use crate::test_util::*;
 
     use super::*;
     use assert_approx_eq::assert_approx_eq;
-    use rstest::{fixture, rstest};
+    use rstest::rstest;
     use rstest_reuse::{self, *};
-    use serde::Deserialize;
     use std::f64::consts::{FRAC_PI_2, FRAC_PI_3, FRAC_PI_4, FRAC_PI_6, FRAC_PI_8, PI, SQRT_2};
-    use std::path::PathBuf;
     use tempfile::NamedTempFile;
-
-    #[derive(Deserialize)]
-    struct PolygonMetadata {
-        area: f64,
-        extreme_points: HashSet<VertexId>,
-        interior_points: HashSet<VertexId>,
-        num_edges: usize,
-        num_triangles: usize,
-        num_vertices: usize,
-    }
-
-    struct PolygonTestCase {
-        polygon: Polygon,
-        metadata: PolygonMetadata,
-    }
-
-    impl PolygonTestCase {
-        fn new(polygon: Polygon, metadata: PolygonMetadata) -> Self {
-            PolygonTestCase { polygon, metadata }
-        }
-    }
-
-    fn load_metadata(name: &str, folder: &str) -> Result<PolygonMetadata, FileError> {
-        let mut path = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-        path.push("polygons");
-        path.push(folder);
-        path.push(format!("{}.meta.json", name));
-        let metadata_str: String = fs::read_to_string(path)?;
-        let metadata = serde_json::from_str(&metadata_str)?;
-        Ok(metadata)
-    }
-
-    #[macro_export]
-    macro_rules! polygon_fixture {
-        ($name:ident, $folder:expr) => {
-            #[fixture]
-            fn $name() -> PolygonTestCase {
-                PolygonTestCase::new(
-                    load_polygon(stringify!($name), stringify!($folder)).unwrap(),
-                    load_metadata(stringify!($name), stringify!($folder)).unwrap(),
-                )
-            }
-        };
-    }
-
-    polygon_fixture!(polygon_1, custom);
-    polygon_fixture!(polygon_2, custom);
-    polygon_fixture!(right_triangle, custom);
-    polygon_fixture!(square_4x4, custom);
-
-    polygon_fixture!(eberly_10, interesting_polygon_archive);
-    polygon_fixture!(eberly_14, interesting_polygon_archive);
-    polygon_fixture!(elgindy_1, interesting_polygon_archive);
-    polygon_fixture!(gray_embroidery, interesting_polygon_archive);
-    polygon_fixture!(held_1, interesting_polygon_archive);
-    polygon_fixture!(held_3, interesting_polygon_archive);
-    polygon_fixture!(held_12, interesting_polygon_archive);
-    polygon_fixture!(held_7a, interesting_polygon_archive);
-    polygon_fixture!(held_7b, interesting_polygon_archive);
-    polygon_fixture!(held_7c, interesting_polygon_archive);
-    polygon_fixture!(held_7d, interesting_polygon_archive);
-    polygon_fixture!(mapbox_building, interesting_polygon_archive);
-    polygon_fixture!(mapbox_dude, interesting_polygon_archive);
-    polygon_fixture!(matisse_alga, interesting_polygon_archive);
-    polygon_fixture!(matisse_blue, interesting_polygon_archive);
-    polygon_fixture!(matisse_icarus, interesting_polygon_archive);
-    polygon_fixture!(matisse_nuit, interesting_polygon_archive);
-    polygon_fixture!(mei_2, interesting_polygon_archive);
-    polygon_fixture!(mei_3, interesting_polygon_archive);
-    polygon_fixture!(mei_4, interesting_polygon_archive);
-    polygon_fixture!(mei_5, interesting_polygon_archive);
-    polygon_fixture!(mei_6, interesting_polygon_archive);
-    polygon_fixture!(meisters_3, interesting_polygon_archive);
-    polygon_fixture!(misc_discobolus, interesting_polygon_archive);
-    polygon_fixture!(misc_fu, interesting_polygon_archive);
-    polygon_fixture!(seidel_3, interesting_polygon_archive);
-    polygon_fixture!(skimage_horse, interesting_polygon_archive);
-    polygon_fixture!(toussaint_1a, interesting_polygon_archive);
-
-    #[template]
-    #[rstest]
-    #[case::right_triangle(right_triangle())]
-    #[case::square_4x4(square_4x4())]
-    #[case::polygon_1(polygon_1())]
-    #[case::polygon_2(polygon_2())]
-    #[case::eberly_10(eberly_10())]
-    #[case::eberly_14(eberly_14())]
-    #[case::elgindy_1(elgindy_1())]
-    #[case::gray_embroidery(gray_embroidery())]
-    #[case::held_1(held_1())]
-    #[case::held_3(held_3())]
-    #[case::held_12(held_12())]
-    #[case::held_7a(held_7a())]
-    #[case::held_7b(held_7b())]
-    #[case::held_7c(held_7c())]
-    #[case::held_7d(held_7d())]
-    #[case::mapbox_building(mapbox_building())]
-    #[case::mapbox_dude(mapbox_dude())]
-    #[case::matisse_alga(matisse_alga())]
-    #[case::matisse_blue(matisse_blue())]
-    #[case::matisse_icarus(matisse_icarus())]
-    #[case::matisse_nuit(matisse_nuit())]
-    #[case::mei_2(mei_2())]
-    #[case::mei_3(mei_3())]
-    #[case::mei_4(mei_4())]
-    #[case::mei_5(mei_5())]
-    #[case::mei_6(mei_6())]
-    #[case::meisters_3(meisters_3())]
-    #[case::misc_discobolus(misc_discobolus())]
-    #[case::misc_fu(misc_fu())]
-    #[case::seidel_3(seidel_3())]
-    #[case::skimage_horse(skimage_horse())]
-    #[case::toussaint_1a(toussaint_1a())]
-    fn all_polygons(#[case] case: PolygonTestCase) {}
-
-    #[template]
-    #[rstest]
-    #[case::polygon_1(polygon_1())]
-    #[case::polygon_2(polygon_2())]
-    #[case::right_triangle(right_triangle())]
-    #[case::square_4x4(square_4x4())]
-    fn all_custom_polygons(#[case] case: PolygonTestCase) {}
-
-    #[template]
-    #[apply(all_custom_polygons)]
-    #[case::eberly_10(eberly_10())]
-    #[case::eberly_14(eberly_14())]
-    #[case::elgindy_1(elgindy_1())]
-    #[case::gray_embroidery(gray_embroidery())]
-    #[case::held_1(held_1())]
-    #[case::held_12(held_12())]
-    #[case::held_3(held_3())]
-    // TODO can add more test cases here, but I've just been going through
-    // and manually verifying they're correct. Some have so many vertices
-    // though it's not practical
-    fn extreme_point_cases(#[case] case: PolygonTestCase) {}
 
 
     #[test]
@@ -891,13 +721,6 @@ mod tests {
             extreme_points.difference(&case.metadata.extreme_points),
             case.metadata.extreme_points.difference(&extreme_points)
         );
-    }
-
-    #[apply(extreme_point_cases)]
-    fn test_convex_hull_from_gift_wrapping(#[case] case: PolygonTestCase) {
-        let hull = case.polygon.convex_hull_from_gift_wrapping();
-        let expected_hull = ConvexHull::new(case.metadata.extreme_points);
-        assert_eq!(hull, expected_hull);
     }
 
     #[apply(extreme_point_cases)]
